# src/main/java/com/testtask_effectivemobile/models/enums/Role.java
package com.testtask_effectivemobile.models.enums;


import org.springframework.security.core.GrantedAuthority;

public enum Role implements GrantedAuthority {
    ROLE_USER,
    ROLE_ADMIN;

    @Override
    public String getAuthority() {
        return name();
    }
}